requirements.txt
django==3.2.6
pillow==8.3.1
djangorestframework==3.12.4
coverage==5.5